+++ src/main/java/cn/yuqi/mini/spring/beans/factory/BeanFactory.java
@@ -1,5 +1,7 @@
 package cn.yuqi.mini.spring.beans.factory;
 
+import cn.yuqi.mini.spring.beans.BeansException;
+
 public interface BeanFactory {
     Object getBean(String beanName) throws BeansException;
 }