+++ src/main/java/cn/yuqi/mini/spring/beans/factory/config/BeanDefinition.java
@@ -1,17 +1,28 @@
 package cn.yuqi.mini.spring.beans.factory.config;
 
+import cn.yuqi.mini.spring.beans.PropertyValues;
+
 public class BeanDefinition {
     private Class<?> beanClass;
     private Object[] constructorArgs;
+    private PropertyValues propertyValues;
 
-    public BeanDefinition(Class<?> beanClass, Object... constructorArgs) {
+    public BeanDefinition(Class<?> beanClass) {
         this.beanClass = beanClass;
+    }
+    public void setConstructorArgs(Object... constructorArgs){
         this.constructorArgs = constructorArgs;
     }
+    public void setPropertyValues(PropertyValues propertyValues){
+        this.propertyValues = propertyValues;
+    }
+
+
     public Class<?> getBeanClass(){
         return this.beanClass;
     }
     public Object[] getConstructorArgs(){
         return this.constructorArgs;
     }
+    public PropertyValues getPropertyValues(){return this.propertyValues;}
 }