+++ src/main/java/cn/yuqi/mini/spring/beans/factory/support/AbstractBeanFactory.java
@@ -1,7 +1,7 @@
 package cn.yuqi.mini.spring.beans.factory.support;
 
 import cn.yuqi.mini.spring.beans.factory.BeanFactory;
-import cn.yuqi.mini.spring.beans.factory.BeansException;
+import cn.yuqi.mini.spring.beans.BeansException;
 import cn.yuqi.mini.spring.beans.factory.config.BeanDefinition;
 
 public abstract class AbstractBeanFactory